changes to use floats

+++ Robust/src/Benchmarks/SingleTM/KMeans/Common.java
@@ -78,7 +78,7 @@ public class Common {
     common_euclidDist2 (float[] pt1, float[] pt2, int numdims)
     {
       int i;
-      float ans = 0.0;
+      float ans = 0.0f;
 
       for (i = 0; i < numdims; i++) {
         ans += (pt1[i] - pt2[i]) * (pt1[i] - pt2[i]);
@@ -101,8 +101,8 @@ public class Common {
       int index = -1;
       int i;
       //double max_dist = FLT_MAX;
-      float max_dist = 3.40282347e+38;
-      float limit = 0.99999;
+      float max_dist = (float)3.40282347e+38;
+      float limit = (float) 0.99999;
 
       /* Find the cluster center id with min distance to pt */
       for (i = 0; i < npts; i++) {
